requirements.txt
fastapi==0.111.0
minio==7.2.7
pinecone-client==4.1.0
openai==1.30.1
tiktoken==0.7.0
requests==2.31.0
PyMuPDF==1.24.3